+++ controls-constructor/src/commonMain/kotlin/space/kscience/controls/constructor/ExternalValueState.kt
@@ -1,29 +1,29 @@
 package space.kscience.controls.constructor
 
-import kotlinx.coroutines.*
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.flow.flow
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
 import space.kscience.controls.time.ValueWithTime
-import kotlin.time.Clock
+import space.kscience.controls.time.clockManager
+import space.kscience.dataforge.context.Context
 import kotlin.time.Duration
 
 
 private open class ExternalValueState<T>(
     val scope: CoroutineScope,
-    val readInterval: Duration,
     initialValue: T,
-    val clock: Clock = Clock.System,
+    val timer: TimerState,
     val reader: suspend () -> T,
 ) : ValueState<T> {
 
-    protected val flow: StateFlow<ValueWithTime<T>> = flow {
-        while (true) {
-            delay(readInterval)
-            emit(ValueWithTime(reader(), clock.now()))
-        }
-    }.stateIn(scope, SharingStarted.Eagerly, ValueWithTime(initialValue, clock.now()))
+    protected val flow: StateFlow<ValueWithTime<T>> = timer.subscribe().map { now ->
+        ValueWithTime(reader(), now)
+    }.stateIn(scope, SharingStarted.Eagerly, ValueWithTime(initialValue, timer.now()))
 
     override val valueWithTime: ValueWithTime<T> get() = flow.value
 
@@ -36,21 +36,26 @@ private open class ExternalValueState<T>(
  * Create a [ValueState] which is constructed by regularly reading external value
  */
 public fun <T> ValueState.Companion.external(
-    scope: CoroutineScope,
+    context: Context,
     readInterval: Duration,
     initialValue: T,
-    clock: Clock = Clock.System,
+    scope: CoroutineScope = context,
     reader: suspend () -> T,
-): ValueState<T> = ExternalValueState(scope, readInterval, initialValue, clock, reader)
+): ValueState<T> = ExternalValueState(scope, initialValue, TimerState(context.clockManager, readInterval), reader)
 
 private class MutableExternalValueState<T>(
-    scope: CoroutineScope,
+    context: Context,
     readInterval: Duration,
     initialValue: T,
-    clock: Clock = Clock.System,
     reader: suspend () -> T,
     val writer: suspend (T) -> Unit,
-) : ExternalValueState<T>(scope, readInterval, initialValue, clock, reader), MutableValueState<T> {
+    scope: CoroutineScope = context,
+) : ExternalValueState<T>(
+    scope = scope,
+    initialValue = initialValue,
+    timer = TimerState(context.clockManager, readInterval),
+    reader = reader
+), MutableValueState<T> {
 
     override var value: T
         get() = super.value
@@ -71,10 +76,10 @@ private class MutableExternalValueState<T>(
  * Create a [MutableValueState] which is constructed by regularly reading external value and allows writing
  */
 public fun <T> ValueState.Companion.external(
-    scope: CoroutineScope,
+    context: Context,
     readInterval: Duration,
     initialValue: T,
-    clock: Clock = Clock.System,
     reader: suspend () -> T,
     writer: suspend (T) -> Unit,
-): MutableValueState<T> = MutableExternalValueState(scope, readInterval, initialValue, clock, reader, writer)
+    scope: CoroutineScope = context,
+): MutableValueState<T> = MutableExternalValueState(context, readInterval, initialValue, reader, writer, scope)