+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Added
+- `TimerState` implements `Clock`.
 - `setCachedValue` API for `CachingDevice` to directly manipulate cached value and solve circular write problem with virtual properties.
 - Flow control simulation
 - Value averaging plot extension